[hive/network]: iptables DPORT target is still in progress.  It's doing what I wanted but I just realized that I need to translate the source package as well.

+++ Tools/neon-cli/Resources/Ubuntu-16.04/source/xt_DPORT/deploy.sh
@@ -1,4 +1,14 @@
-﻿# This script builds and deploys the xt_DPORT iptables modules as required.
+﻿#!/bin/bash
+#------------------------------------------------------------------------------
+# FILE:         deploy.sh
+# CONTRIBUTOR:  Jeff Lill
+# COPYRIGHT:    Copyright (c) 2016-2018 by neonFORGE, LLC.  All rights reserved.
+#
+# This script builds and deploys the xt_DPORT iptables modules as required.
+
+# Make sure that we have loaded the standard environment variables.
+
+. /etc/neon/hive.conf.sh
 
 # $hack(jeff.lill):
 #
@@ -17,31 +27,68 @@
 # The only real assumption here is that the host node is rebooted after
 # upgrading the Linux kernel.  [neon hive upgrade linux] does this.
 
-# Change to the directory hosting this script.
+# We need to rebuild the modules after every reboot in case the 
+# Linux kernel was upgraded.  We're going to use the presence of
+# the [xt-DROP-built] file in a tmpfs to indicate that the module
+# has been built already.  Since this is on a RAM drive, the file
+# will no longer exist after a reboot, indicating that the modules
+# need to be rebuilt.
 
-pushd "$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
+mkdir -p ${NEON_TMPFS_FOLDER}
 
-# Build and install userspace module.
+if [ ! -f ${NEON_TMPFS_FOLDER}/xt-DROP-built ] ; then
 
-cp Makefile-so Makefile
-make libxt_DPORT.so
-cp libxt_DPORT.so /lib/xtables
-chmod 644 /lib/xtables/libxt_DPORT.so
-rm Makefile
+    echo "[INFO] Building iptables DPORT target extension."
 
-# Build and install the kernel module.
+    # Change to the directory hosting this script.
 
-cp Makefile-ko Makefile
-make
-cp xt_DPORT.ko /lib/modules/`uname -r`/kernel/net/netfilter
-rm Makefile
+    pushd "$( cd "$( dirname "${BASH_SOURCE[0]}" )" &>/dev/null && pwd )"
+
+    # Build and install the userspace module.
+
+    cp Makefile-so Makefile
+    
+    if [ -f libxt_DPORT.so ] ; then
+        rm libxt_DPORT.so
+    fi
+    
+    make libxt_DPORT.so
+    cp libxt_DPORT.so /lib/xtables/libxt_DPORT.so
+    chmod 644 /lib/xtables/libxt_DPORT.so
+    rm Makefile
+
+    # Build and install the kernel module.
+
+    cp Makefile-ko Makefile
+    
+    if [ -f xt_DPORT.ko ] ; then
+        rm xt_DPORT.ko
+    fi
+
+    make
+    cp xt_DPORT.ko /lib/modules/`uname -r`/kernel/net/netfilter/xt_DPORT.ko
+    chmod 644 /lib/modules/`uname -r`/kernel/net/netfilter/xt_DPORT.ko
+    rm Makefile
+
+    # Indicate that the modules were built or at least we tried
+    # to build them.  We're not going to try again.
+
+    touch ${NEON_TMPFS_FOLDER}/xt-DROP-built
+
+    # Restore the directory.
+
+    popd
+fi
 
 # Load the kernel module if it's not already loaded.
 
 if ! grep xt_DPORT /proc/modules &> /dev/nul ; then
+
+    echo "[INFO] Loading iptables DPORT target extension."
+
+    # Ensure that iptables is loaded first by listing some rules
+    # and then load our module.
+
+    iptables -S PREROUTING -t raw &> /dev/nul
     insmod /lib/modules/`uname -r`/kernel/net/netfilter/xt_DPORT.ko
 fi
-
-# Restore the directory.
-
-popd
